+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enforce hard space before `\ref`.
 - Check operator names inside math mode.
 - Enforce `\textmu` instead of `\mu` in math mode.
+- Enforce `\left` and `\right` for parentheses in math mode.
 
 ### Fixed